requirements.txt
Django==1.8.4
Pillow==3.4.2
Shapely==1.5.17
django-appconf==1.0.1
django-compressor==1.5
django-countries==3.3
django-fiber==1.4
django-geojson==2.8.1
django-leaflet==0.16.0
django-mptt==0.7.4
django-olwidget==0.61.0
django-tastypie==0.12.2
djangorestframework==2.4.8
easy-thumbnails==2.2
fastkml==0.11
lxml==3.7.3
psycopg2==2.7.1
pygeoif==0.6
pyshp==1.2.3
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
unicodecsv==0.14.1
utm==0.4.0